1. Add context folder selection at start for git workflow 2. Check for component folder already existing
